+++ cypress.config.js
@@ -4,7 +4,7 @@ module.exports = defineConfig({
   viewportWidth: 500,
   viewportHeight: 600,
   video: false,
-  defaultCommandTimeout: 16000, // E2E Edge test is slow in CI/CD
+  defaultCommandTimeout: 8000, // E2E Edge test is slow in CI/CD
   e2e: {
     baseUrl: 'http://localhost:8080/popup',
   },